clean: created CreateTimestampableTrait and updated Entities that have a createdAt field to use it.
